ANGELGUARD/requirements.txt
psutil
PyQt5
requests
python-dotenv
openai